Replace reference data to make repo internally compatible
This commit replaces the reference tar with individual reference files by slicing the genome down to limited interval sets. The final reduction of file size is from around 17mb compressed tar to 52k uncompressed files which are ready to use.

In addition, it switches to using a BED file for intervals for easier compatibility with other tools.